Module recursion::recursive_tree

Re-exports

pub use crate::recursive_tree::arena_eval::ArenaIndex;

pub use crate::recursive_tree::stack_machine_eval::StackMarker;

Modules

arena_eval

Recursive structure that uses an arena to quickly collapse recursive structures.

stack_machine_eval

Recursive structure stored using a compact stack machine representation Collapsed via stack machine evaluation.

Structs

RecursiveTree

A recursive structure with layers of partially-applied type Layer, where Index is the type that Layer is parameterized over and Wrapped is Layer<Index>

RecursiveTreeRef

A reference to some recursive structure with layers of partially-applied type Layer, where Index is the type that Layer is parameterized over and Wrapped is Layer<Index>