Trait recursion::map_layer::MapLayerRef

pub trait MapLayerRef<B> {
    type Unwrapped;
    type To;

    fn map_layer_ref<F: FnMut(Self::Unwrapped) -> B>(&self, f: F) -> Self::To;
}

Required Associated Types

type Unwrapped

type To

Required Methods

fn map_layer_ref<F: FnMut(Self::Unwrapped) -> B>(&self, f: F) -> Self::To

Additional constraint not present in haskell, req’d for stack machine eval: for any F, F, etc, where all structure but B/A is identical, fmap must visit nodes in the same order each time it is called given that F is created by mapping some function over F note that enforcing this property may be problematic for, Hashmaps/Sets/etc